requirements.txt
spotipy
mysql-connector-python